// src/providers/types/api-private.ts
export type FaultAction = 'APPROVED'|'REJECTED'|'ERASED'|'RESTAB_WAITING'|'RESTAB_PENDING'|'RESTABLISHED'|'RESTAB_REJECTED'|'PENDING';